% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacket-package.R
\docType{package}
\name{jacket-package}
\alias{jacket}
\alias{jacket-package}
\title{jacket: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
