% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacket_dt.R
\name{jacket_dt}
\alias{jacket_dt}
\title{DT for Corcept}
\usage{
jacket_dt(x, cols)
}
\arguments{
\item{x}{a data frame or .webpro_comparison object (created from `webpro_compare()`)}
}
\description{
Wrap a DT object with some additional functionality specific to Corcept
}
