% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas.R
\name{read_sas}
\alias{read_sas}
\title{Read SAS}
\usage{
read_sas(..., n_max = Inf)
}
\arguments{
\item{...}{strings to be pasted together to form a path (via `fs::path(...)`)}
}
\description{
Read a SAS file
}
